Moved textToXML method to export utilities file
